Ensure active_model_serializers is required
I ran into a bug in an including project where the Client class didn't
have access to ActiveModel::Serializers. I realized that the spec helper
was corrupting the test by including that file for all tests. Oops. No
one appeared to be using the CustomSerializer class so I remove it here.